non-blocking send, warn past deadline, m2 refactor, m1.1 new trip=20us, 20x20 nodes

+++ event.go
@@ -27,7 +27,7 @@ type TimedUcastEvent struct {
 }
 
 func newTimedUcastEvent(src RunnerInterface, when time.Duration, tgt RunnerInterface) *TimedUcastEvent {
-	assert(int64(when) > 0)
+	assert(when > 0)
 	triggertime := Now.Add(when)
 	return &TimedUcastEvent{
 		crtime: Now,
